refactor(torghut): remove proof floor shared context
