+++ services/torghut/app/api/proof_floor_payloads/evidence_clock_payloads.py
@@ -1,0 +1,52 @@
+"""Evidence-clock payload assembly."""
+
+from __future__ import annotations
+
+from collections.abc import Mapping
+from typing import Any
+
+from app.config import settings
+from app.trading.evidence_clock_arbiter import build_evidence_clock_arbiter_and_exchange
+
+from .status_refs import build_torghut_stage_clearance_packet_ref_payload
+
+
+def build_evidence_clock_payloads(
+    *,
+    torghut_revision: str | None,
+    dependency_quorum: Mapping[str, Any],
+    hypothesis_payload: Mapping[str, Any],
+    quant_evidence: Mapping[str, Any],
+    market_context_status: Mapping[str, Any],
+    tca_summary: Mapping[str, Any],
+    empirical_jobs_status: Mapping[str, Any],
+    proof_floor: Mapping[str, Any],
+    routeability_repair_acceptance_ledger: Mapping[str, Any],
+    profit_signal_quorum: Mapping[str, Any],
+    live_submission_gate: Mapping[str, Any],
+    build: Mapping[str, Any],
+    clickhouse_ta_status: Mapping[str, Any],
+) -> tuple[dict[str, object], dict[str, object]]:
+    return build_evidence_clock_arbiter_and_exchange(
+        account_label=settings.trading_account_label,
+        window=settings.trading_jangar_quant_window,
+        trading_mode=settings.trading_mode,
+        torghut_revision=torghut_revision,
+        build=build,
+        hypothesis_payload=hypothesis_payload,
+        quant_evidence=quant_evidence,
+        market_context_status=market_context_status,
+        tca_summary=tca_summary,
+        empirical_jobs_status=empirical_jobs_status,
+        proof_floor_receipt=proof_floor,
+        routeability_repair_acceptance_ledger=routeability_repair_acceptance_ledger,
+        profit_signal_quorum=profit_signal_quorum,
+        live_submission_gate=live_submission_gate,
+        torghut_custody_ref=build_torghut_stage_clearance_packet_ref_payload(
+            dependency_quorum
+        ),
+        clickhouse_ta_status=clickhouse_ta_status,
+    )
+
+
+__all__ = ("build_evidence_clock_payloads",)
