+++ services/torghut/app/api/proof_floor_payloads/proof_floor_receipts.py
@@ -1,0 +1,108 @@
+"""Proof-floor and route-reacquisition payload assembly."""
+
+from __future__ import annotations
+
+from collections.abc import Mapping
+from typing import Any, cast
+
+from app.api.proof_contracts import PROFITABILITY_PROOF_FLOOR_TCA_MAX_AGE_SECONDS
+from app.config import settings
+from app.trading.proof_floor import build_profitability_proof_floor_receipt
+from app.trading.renewal_bond_profit_escrow import build_renewal_bond_profit_escrow
+from app.trading.route_reacquisition_board import build_route_reacquisition_board
+
+from .paper_route_probe_targets import bounded_paper_route_probe_target_symbols
+from .status_refs import (
+    build_simple_lane_status_payload,
+    route_continuity_packet_for_proof_floor,
+)
+
+
+def build_profitability_proof_floor_payload(
+    *,
+    state: object,
+    torghut_revision: str | None,
+    live_submission_gate: Mapping[str, Any],
+    hypothesis_payload: Mapping[str, Any],
+    empirical_jobs_status: Mapping[str, Any],
+    quant_evidence: Mapping[str, Any],
+    market_context_status: Mapping[str, Any],
+    tca_summary: Mapping[str, Any],
+    simple_lane_status: Mapping[str, Any] | None = None,
+) -> dict[str, object]:
+    return build_profitability_proof_floor_receipt(
+        account_label=settings.trading_account_label,
+        torghut_revision=torghut_revision,
+        trading_mode=settings.trading_mode,
+        market_session_open=cast(
+            bool | None,
+            getattr(state, "market_session_open", None),
+        ),
+        live_submission_gate=live_submission_gate,
+        hypothesis_payload=hypothesis_payload,
+        empirical_jobs_status=empirical_jobs_status,
+        quant_evidence=quant_evidence,
+        market_context_status=market_context_status,
+        tca_summary=tca_summary,
+        simple_lane_status=simple_lane_status or build_simple_lane_status_payload(),
+        paper_route_probe_target_symbols=bounded_paper_route_probe_target_symbols(
+            live_submission_gate
+        ),
+        tca_max_age_seconds=PROFITABILITY_PROOF_FLOOR_TCA_MAX_AGE_SECONDS,
+    )
+
+
+def build_renewal_bond_profit_escrow_payload(
+    *,
+    state: object,
+    torghut_revision: str | None,
+    dependency_quorum: Mapping[str, Any],
+    live_submission_gate: Mapping[str, Any],
+    proof_floor: Mapping[str, Any],
+    hypothesis_payload: Mapping[str, Any],
+    empirical_jobs_status: Mapping[str, Any],
+    quant_evidence: Mapping[str, Any],
+    market_context_status: Mapping[str, Any],
+    tca_summary: Mapping[str, Any],
+) -> dict[str, object]:
+    return build_renewal_bond_profit_escrow(
+        account_label=settings.trading_account_label,
+        torghut_revision=torghut_revision,
+        trading_mode=settings.trading_mode,
+        market_session_open=cast(
+            bool | None,
+            getattr(state, "market_session_open", None),
+        ),
+        jangar_dependency_quorum=dependency_quorum,
+        live_submission_gate=live_submission_gate,
+        proof_floor=proof_floor,
+        hypothesis_payload=hypothesis_payload,
+        empirical_jobs_status=empirical_jobs_status,
+        quant_evidence=quant_evidence,
+        market_context_status=market_context_status,
+        tca_summary=tca_summary,
+        tca_max_age_seconds=PROFITABILITY_PROOF_FLOOR_TCA_MAX_AGE_SECONDS,
+    )
+
+
+def build_route_reacquisition_board_payload(
+    *,
+    proof_floor: Mapping[str, Any],
+    active_revision: str | None,
+) -> dict[str, object]:
+    return build_route_reacquisition_board(
+        proof_floor_receipt=proof_floor,
+        route_reacquisition_book=cast(
+            Mapping[str, Any] | None,
+            proof_floor.get("route_reacquisition_book"),
+        ),
+        active_revision=active_revision,
+        jangar_continuity=route_continuity_packet_for_proof_floor(proof_floor),
+    )
+
+
+__all__ = (
+    "build_profitability_proof_floor_payload",
+    "build_renewal_bond_profit_escrow_payload",
+    "build_route_reacquisition_board_payload",
+)
